third_party/WebKit/LayoutTests/platform/win/virtual/threaded/inspector/tracing/timeline-js/compile-script-expected.txt

CONSOLE MESSAGE: line 1: SCRIPT TAG
Tests the Timeline instrumentation for CompileScript event.

v8.compile Properties:
{
    data : {
        columnNumber : <number>
        lineNumber : <number>
        url : 
    }
    endTime : <number>
    startTime : <number>
    type : "v8.compile"
}
Text details for v8.compile: undefined
